feat: variations and catalog type fixes (#626)
* feat: added MatrixObject for variation matrix

* feat: Includes values for catalog resources

Created a more constrained catalog resource type that can be used to
define any includes that are applicable to the Catalog endpoints but
maybe not other endpoints that are dependent on the Resource interface.

* feat: meta data defined for product responses

Including a more constrained version of Product Variations type
definition. Omiting properties that are not part of the
simplified response.

* refactor: style match

* feat: parent relationship typed

* fix: product response meta is on all responses

Nested properties are optional as they don't show on all responses.